Change Error handling of JenkinsApi
This PR should address the issue where it is not 100% clear whether
fetching of a JenkinsBuild itself failed (Jenkins unreachable?)
or if the jenkins build just doesn't exist.

Therefore, just throw the checked exception in case of failure,
while null signals that that build does not exist.

This allows proper handling in both the VersionChecker (no stacktraces
logged), as well as other methods (like JarCommand) that depend on it.