Update the Launcher system
Previously, due to the main process the console created being able to exit,
the console returned back to the shell, even if the bot was restarted after.

To prevent that, the launcher will now always run, wait for bot to exit,
check return code and update/restart if applicable.

A side effect of this change is, that the Launcher itself
can currently not be updated anymore.